+++ paddle/fluid/framework/new_executor/interpretercore.h
@@ -12,55 +12,40 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 #pragma once
-
-#include <map>
-#include <queue>
-#include <string>
-#include <unordered_map>
-#include <vector>
-
-#include "paddle/fluid/framework/details/exception_holder.h"
-#include "paddle/fluid/framework/new_executor/garbage_collector/garbage_collector.h"
-#include "paddle/fluid/framework/new_executor/interpreter/dependency_builder.h"
-#include "paddle/fluid/framework/new_executor/interpreter/execution_config.h"
-#include "paddle/fluid/framework/new_executor/interpreter/interpreter_util.h"
-#include "paddle/fluid/framework/new_executor/interpreter/stream_analyzer.h"
-#include "paddle/fluid/framework/new_executor/new_executor_defs.h"
-#include "paddle/fluid/framework/new_executor/profiler.h"
-#include "paddle/fluid/framework/program_desc.h"
-#include "paddle/fluid/framework/tensor.h"
-#include "paddle/fluid/framework/variable.h"
-#include "paddle/fluid/memory/allocation/spin_lock.h"
-#include "paddle/fluid/platform/device_event.h"
-#include "paddle/ir/core/program.h"
-#include "paddle/ir/core/value.h"
+#include "paddle/fluid/framework/new_executor/interpreter_base_impl.h"
 
 DECLARE_bool(new_executor_use_local_scope);
+
+namespace ir {
+class Program;
+}  // namespace ir
+
+namespace ir {
+class Program;
+}  // namespace ir
 
 namespace paddle {
 namespace framework {
 
+class InterpreterBaseImpl;
+
 class InterpreterCore {
   using ExecutionConfig = interpreter::ExecutionConfig;
-  using InstructionSchedulingPriorityLess = std::function<bool(size_t, size_t)>;
-  using SchedulingQueue =
-      std::priority_queue<size_t,
-                          std::vector<size_t>,
-                          InstructionSchedulingPriorityLess>;
+  using HookFunc = std::function<void(OperatorBase*, Scope*)>;
 
  public:
   InterpreterCore(const platform::Place& place,
                   const BlockDesc& block,
                   Scope* scope,
                   const ExecutionConfig& execution_config = ExecutionConfig());
-
+  // This constructor is for New IR.
   InterpreterCore(const platform::Place& place,
-                  const BlockDesc& block,
-                  Scope* scope,
                   std::unique_ptr<::ir::Program> ir_prog,
+                  Scope* scope,
                   const ExecutionConfig& execution_config = ExecutionConfig());
-
   ~InterpreterCore();
+
+  const InterpreterBaseImpl* Impl() const { return impl_.get(); }
 
   paddle::framework::FetchList Run(
       const std::vector<std::string>& feed_names,
@@ -83,127 +68,14 @@ class InterpreterCore {
 
   void reset_scope(Scope* new_scope);
 
-  const platform::Place& GetPlace() const { return place_; }
+  const platform::Place& GetPlace() const;
 
-  using HookFunc = std::function<void(OperatorBase*, Scope*)>;
-  void SetOutputHooks(const std::vector<HookFunc>& hookfuncs) {
-    hookfuncs_ = hookfuncs;
-  }
+  void SetOutputHooks(const std::vector<HookFunc>& hookfuncs);
 
  private:
   DISABLE_COPY_AND_ASSIGN(InterpreterCore);
-  // build graph
-  void Convert(std::vector<paddle::framework::OpFuncNode>* op_func_nodes);
-  void BuildOperatorDependences();
-  void BuildAndCacheInstructionCtx(Instruction* instr_node);
-  void BuildSkipShareLoDInfo();
-  void UpdateSyncOpNum();
-  void AnalyseExecuteOrderForTrace();
 
-  // inplace
-  void BuildInplace();
-  bool BuildInplaceCheckVarIsOnlyInput(
-      const std::vector<std::vector<size_t>>& input_var2op, size_t var_index);
-  void SetFeedVarsInplaceSkip(const std::vector<std::string>& feed_names);
-
-  // cuda graph
-  void CheckCUDAGraphBeforeRun(const std::vector<std::string>& feed_names);
-  void PrepareForCUDAGraphCapture();
-
-  // execution
-  void RunImpl();
-  void ExecuteInstructionList(const std::vector<Instruction>& vec_instr);
-  void RunInstructionAsync(size_t instr_id);
-  void RunInstruction(const Instruction& instr_node);
-  void RunNextInstructions(const Instruction& instr_id,
-                           SchedulingQueue* reserved_next_ops);
-  void RunOperator(const Instruction& instr_node);
-  // Trace
-  void TraceInstructionList(const std::vector<Instruction>& vec_instr);
-
-  // only used when program contains no feed op
-  void Prepare(const std::vector<std::string>& feed_names,
-               const std::vector<phi::DenseTensor>& feed_tensors,
-               bool prepare_feed);
-
-  void RecordMemcpyD2H(const Instruction& instr_node);
-
-  // gc
-  void RecordStreamForGC(const Instruction& instr);
-  void CheckGC(const Instruction& instr);
-  void ClearLoDTensorArrayInLocalScope();
-
-  // workqueue
-  std::shared_ptr<interpreter::AsyncWorkQueue> GetWorkQueue();
-
-  // scope
-  bool HasLocalScope() const;
-
-  // For log and debug
-  std::string GetDepsString() const;
-
- private:
-  bool is_build_{false};
-  bool static_build_{false};
-
-  const platform::Place place_;
-  const BlockDesc& block_;  // not owned
-
-  interpreter::DependencyBuilder dependency_builder_;
-  interpreter::StreamAnalyzer stream_analyzer_;
-
-  // NOTE(zhiqiu): when add fetch ops in GetInterpreterCore, we will
-  // copy a new program and block, the copy_program_ here is used to
-  // hold the program, otherwise block_ maybe not valid after the
-  // new program is deleted.
-  std::shared_ptr<ProgramDesc> copy_program_{nullptr};
-
-  // from variable scope
-  std::vector<Variable*> var_list_;
-  std::map<std::string, int> name2id_;
-  std::vector<VariableMetaInfo> vec_meta_info_;
-
-  std::vector<Instruction> vec_instruction_;  // deconstruct before OpFuncNode
-
-  std::atomic<size_t> unfinished_op_number_{0};
-
-  ExecutionConfig execution_config_;
-
-  VariableScope var_scope_;
-  Scope* local_scope_{nullptr};  // not owned
-
-  EventsWaiter main_thread_blocker_;
-  std::shared_ptr<interpreter::AsyncWorkQueue> async_work_queue_;
-
-  details::ExceptionHolder exception_holder_;
-  std::shared_ptr<EventsWaiter::EventNotifier> exception_notifier_{nullptr};
-  std::shared_ptr<EventsWaiter::EventNotifier> completion_notifier_{nullptr};
-
-  std::unique_ptr<InterpreterCoreGarbageCollector> gc_;
-
-  // last_live_ops_[i] contains the id of operators that last access the i-th
-  // var
-  std::map<size_t, std::set<size_t>> last_live_ops_;
-
-  // dependecy_count_[i] contains the number of dependencies that the i-th op
-  // need to wait
-  std::vector<size_t> dependecy_count_;
-
-  std::vector<std::shared_ptr<interpreter::OpDepInfo>> deps_;
-  std::vector<std::shared_ptr<interpreter::VarRefInfo>> refs_;
-
-  // used for Trace
-  int64_t sync_op_num_{-1};
-  std::vector<size_t> trace_execute_order_;
-
-  InstructionSchedulingPriorityLess instruction_scheduling_priority_less;
-
-  std::vector<HookFunc> hookfuncs_;
-
-  // The next only for new IR
-  std::unique_ptr<::ir::Program> ir_program_{nullptr};
-
-  std::unordered_map<::ir::Value, std::string> value_2_var_name_map_;
+  std::unique_ptr<InterpreterBaseImpl> impl_;
 };
 
 }  // namespace framework
