+++ test/cpp/new_executor/standalone_executor_new_ir_test.cc
@@ -69,8 +69,7 @@ TEST(StandaloneExecutor, run) {
   Scope scope;
 
   ProgramDesc prog_desc;
-  InterpreterCore test_core(
-      place, prog_desc.Block(0), &scope, std::move(kernel_program));
+  InterpreterCore test_core(place, std::move(kernel_program), &scope);
 
   test_core.Run({});
